implemented @metaclips review
