+++ fyne/pages/handler/historypagehandler/txdetailscomponent.go
@@ -1,0 +1,242 @@
+package historypagehandler
+
+import (
+	"fmt"
+	"net/url"
+	"strconv"
+	"time"
+
+	"fyne.io/fyne"
+	"fyne.io/fyne/canvas"
+	"fyne.io/fyne/layout"
+	"fyne.io/fyne/theme"
+	"fyne.io/fyne/widget"
+
+	"github.com/decred/dcrd/chaincfg/chainhash"
+	"github.com/decred/dcrd/dcrutil"
+	"github.com/raedahgroup/dcrlibwallet"
+	"github.com/raedahgroup/godcr/fyne/assets"
+	"github.com/raedahgroup/godcr/fyne/helpers"
+	"github.com/raedahgroup/godcr/fyne/pages/handler/values"
+	"github.com/raedahgroup/godcr/fyne/widgets"
+)
+
+func (historyPage *HistoryPageData) fetchTxDetails(hash string) {
+	messageLabel := widget.NewLabelWithStyle("Fetching data..", fyne.TextAlignCenter, fyne.TextStyle{})
+	time.AfterFunc(time.Millisecond*300, func() {
+		if historyPage.TabMenu.CurrentTabIndex() == 1 {
+			messageLabel.Hide()
+		}
+	})
+
+	txDetailslabel := widget.NewLabelWithStyle("Transaction Details", fyne.TextAlignLeading, fyne.TextStyle{Bold: true, Italic: true})
+
+	var txDetailsPopUp *widget.PopUp
+	minimizeIcon := widgets.NewImageButton(theme.CancelIcon(), nil, func() { txDetailsPopUp.Hide() })
+	errorMessageLabel := widget.NewLabelWithStyle("", fyne.TextAlignCenter, fyne.TextStyle{Bold: true})
+
+	txDetailsErrorMethod := func() {
+		txErrorDetailsOutput := widget.NewVBox(
+			widget.NewHBox(
+				txDetailslabel,
+				widgets.NewHSpacer(txDetailslabel.MinSize().Width*3),
+				minimizeIcon,
+			),
+			widget.NewHBox(errorMessageLabel),
+		)
+		txDetailsPopUp = widget.NewModalPopUp(widget.NewVBox(fyne.NewContainer(txErrorDetailsOutput)), historyPage.Window.Canvas())
+		txDetailsPopUp.Show()
+	}
+
+	chainHash, err := chainhash.NewHashFromStr(hash)
+	if err != nil {
+		helpers.ErrorHandler(fmt.Sprintf("Error fetching generating chainhash from for \n %s \n %s ", hash, err.Error()), errorMessageLabel)
+		txDetailsErrorMethod()
+		return
+	}
+
+	txDetails, err := historyPage.MultiWallet.WalletWithID(historyPage.selectedWalletID).GetTransactionRaw(chainHash[:])
+	if err != nil {
+		helpers.ErrorHandler(fmt.Sprintf("Error fetching transaction details for \n %s \n %s ", hash, err.Error()), errorMessageLabel)
+		txDetailsErrorMethod()
+		return
+	}
+
+	var confirmations int32 = 0
+	if txDetails.BlockHeight != -1 {
+		confirmations = historyPage.MultiWallet.WalletWithID(historyPage.selectedWalletID).GetBestBlock() - txDetails.BlockHeight + 1
+	}
+
+	var status string
+	var spendUnconfirmed = historyPage.MultiWallet.ReadBoolConfigValueForKey(dcrlibwallet.SpendUnconfirmedConfigKey, true)
+	if spendUnconfirmed || confirmations > dcrlibwallet.DefaultRequiredConfirmations {
+		status = "Confirmed"
+	} else {
+		status = "Pending"
+	}
+
+	copyAbleText := func(text string, copyAble bool) *widgets.ClickableBox {
+		var textToCopy *canvas.Text
+		if copyAble {
+			textToCopy = canvas.NewText(text, values.Blue)
+		} else {
+			textToCopy = canvas.NewText(text, values.DefaultTextColor)
+		}
+		textToCopy.TextSize = 14
+		textToCopy.Alignment = fyne.TextAlignTrailing
+
+		return widgets.NewClickableBox(widget.NewHBox(textToCopy),
+			func() {
+				messageLabel.SetText("Data Copied")
+				clipboard := historyPage.Window.Clipboard()
+				clipboard.SetContent(text)
+				messageLabel.Show()
+
+				time.AfterFunc(time.Second*2, func() {
+					if historyPage.TabMenu.CurrentTabIndex() == 1 {
+						messageLabel.Hide()
+					}
+				})
+			},
+		)
+	}
+
+	tableConfirmations := widget.NewHBox(
+		widget.NewLabelWithStyle("Confirmations:", fyne.TextAlignTrailing, fyne.TextStyle{Bold: true}),
+		widget.NewLabelWithStyle(strconv.Itoa(int(confirmations)), fyne.TextAlignCenter, fyne.TextStyle{}),
+	)
+
+	tableHash := widget.NewHBox(
+		widget.NewLabelWithStyle("Transaction ID:", fyne.TextAlignTrailing, fyne.TextStyle{Bold: true}),
+		copyAbleText(txDetails.Hash, true),
+	)
+
+	tableBlockHeight := widget.NewHBox(
+		widget.NewLabelWithStyle("Block Height:", fyne.TextAlignTrailing, fyne.TextStyle{Bold: true}),
+		widget.NewLabelWithStyle(strconv.Itoa(int(txDetails.BlockHeight)), fyne.TextAlignCenter, fyne.TextStyle{}),
+	)
+	tableDirection := widget.NewHBox(
+		widget.NewLabelWithStyle("Direction:", fyne.TextAlignTrailing, fyne.TextStyle{Bold: true}),
+		widget.NewLabelWithStyle(dcrlibwallet.TransactionDirectionName(txDetails.Direction), fyne.TextAlignCenter, fyne.TextStyle{}),
+	)
+	tableType := widget.NewHBox(
+		widget.NewLabelWithStyle("Type:", fyne.TextAlignTrailing, fyne.TextStyle{Bold: true}),
+		widget.NewLabelWithStyle(txDetails.Type, fyne.TextAlignCenter, fyne.TextStyle{}),
+	)
+	tableAmount := widget.NewHBox(
+		widget.NewLabelWithStyle("Amount:", fyne.TextAlignTrailing, fyne.TextStyle{Bold: true}),
+		widget.NewLabelWithStyle(dcrutil.Amount(txDetails.Amount).String(), fyne.TextAlignCenter, fyne.TextStyle{}),
+	)
+	tableSize := widget.NewHBox(
+		widget.NewLabelWithStyle("Size:", fyne.TextAlignTrailing, fyne.TextStyle{Bold: true}),
+		widget.NewLabelWithStyle(strconv.Itoa(txDetails.Size)+" Bytes", fyne.TextAlignCenter, fyne.TextStyle{}),
+	)
+	tableFee := widget.NewHBox(
+		widget.NewLabelWithStyle("Fee:", fyne.TextAlignTrailing, fyne.TextStyle{Bold: true}),
+		widget.NewLabelWithStyle(dcrutil.Amount(txDetails.Fee).String(), fyne.TextAlignCenter, fyne.TextStyle{}),
+	)
+	tableFeeRate := widget.NewHBox(
+		widget.NewLabelWithStyle("Fee Rate:", fyne.TextAlignTrailing, fyne.TextStyle{Bold: true}),
+		widget.NewLabelWithStyle(dcrutil.Amount(txDetails.FeeRate).String(), fyne.TextAlignCenter, fyne.TextStyle{}),
+	)
+	tableStatus := widget.NewHBox(
+		widget.NewLabelWithStyle("Status:", fyne.TextAlignTrailing, fyne.TextStyle{Bold: true}),
+		widget.NewLabelWithStyle(status, fyne.TextAlignCenter, fyne.TextStyle{}),
+	)
+	tableDate := widget.NewHBox(
+		widget.NewLabelWithStyle("Date:", fyne.TextAlignTrailing, fyne.TextStyle{Bold: true}),
+		widget.NewLabelWithStyle(fmt.Sprintf("%s UTC", dcrlibwallet.FormatUTCTime(txDetails.Timestamp)), fyne.TextAlignCenter, fyne.TextStyle{}),
+	)
+
+	var txInput widgets.Table
+	inputTableColumnLabels := widget.NewHBox(
+		widget.NewLabelWithStyle("Previous Outpoint", fyne.TextAlignLeading, fyne.TextStyle{Bold: true}),
+		widget.NewLabelWithStyle("Account", fyne.TextAlignCenter, fyne.TextStyle{Bold: true}),
+		widget.NewLabelWithStyle("Amount", fyne.TextAlignCenter, fyne.TextStyle{Bold: true}))
+
+	var inputBox []*widget.Box
+	for i := range txDetails.Inputs {
+		inputBox = append(inputBox, widget.NewHBox(
+			copyAbleText(txDetails.Inputs[i].PreviousOutpoint, true),
+			copyAbleText(txDetails.Inputs[i].AccountName, false),
+			copyAbleText(dcrutil.Amount(txDetails.Inputs[i].Amount).String(), false),
+		))
+	}
+	txInput.NewTable(inputTableColumnLabels, inputBox...)
+
+	var txOutput widgets.Table
+	outputTableColumnLabels := widget.NewHBox(
+		widget.NewLabelWithStyle("Address", fyne.TextAlignCenter, fyne.TextStyle{Bold: true}),
+		widget.NewLabelWithStyle("Account", fyne.TextAlignCenter, fyne.TextStyle{Bold: true}),
+		widget.NewLabelWithStyle("Value", fyne.TextAlignCenter, fyne.TextStyle{Bold: true}),
+		widget.NewLabelWithStyle("Type", fyne.TextAlignCenter, fyne.TextStyle{Bold: true}))
+
+	var outputBox []*widget.Box
+	for i := range txDetails.Outputs {
+		outputBox = append(outputBox, widget.NewHBox(
+			copyAbleText(txDetails.Outputs[i].AccountName, false),
+			copyAbleText(txDetails.Outputs[i].Address, true),
+			copyAbleText(dcrutil.Amount(txDetails.Outputs[i].Amount).String(), false),
+			copyAbleText(txDetails.Outputs[i].ScriptType, false),
+		))
+	}
+	txOutput.NewTable(outputTableColumnLabels, outputBox...)
+
+	tableData := widget.NewVBox(
+		tableConfirmations,
+		tableHash,
+		tableBlockHeight,
+		tableDirection,
+		tableType,
+		tableAmount,
+		tableSize,
+		tableFee,
+		tableFeeRate,
+		tableStatus,
+		tableDate,
+	)
+
+	link, err := url.Parse(fmt.Sprintf("https://%s.dcrdata.org/tx/%s", values.NetType, txDetails.Hash))
+	if err != nil {
+		helpers.ErrorHandler(fmt.Sprintf("Error: ", err.Error()), errorMessageLabel)
+		txDetailsErrorMethod()
+		return
+	}
+
+	redirectWidget := widget.NewHBox(
+		widget.NewHyperlinkWithStyle("View on dcrdata", link, fyne.TextAlignLeading, fyne.TextStyle{Bold: true}),
+		widgets.NewHSpacer(20),
+		widget.NewIcon(historyPage.icons[assets.RedirectIcon]),
+	)
+
+	txDetailsData := widget.NewVBox(
+		widgets.NewHSpacer(10),
+		tableData,
+		canvas.NewLine(values.TxdetailsLineColor),
+		redirectWidget,
+		widgets.NewHSpacer(10),
+		canvas.NewLine(values.TxdetailsLineColor),
+		widget.NewLabelWithStyle("Inputs", fyne.TextAlignLeading, fyne.TextStyle{Bold: true}),
+		txInput.Result,
+		widgets.NewHSpacer(10),
+		canvas.NewLine(values.TxdetailsLineColor),
+		widget.NewLabelWithStyle("Outputs", fyne.TextAlignLeading, fyne.TextStyle{Bold: true}),
+		txOutput.Result,
+		widgets.NewHSpacer(10),
+	)
+
+	txDetailsScrollContainer := widget.NewScrollContainer(txDetailsData)
+	txDetailsOutput := widget.NewVBox(
+		widgets.NewHSpacer(10),
+		widget.NewHBox(
+			txDetailslabel,
+			widgets.NewHSpacer(txDetailsData.MinSize().Width-200),
+			minimizeIcon,
+		),
+		widget.NewHBox(widgets.NewHSpacer(txDetailsScrollContainer.MinSize().Width*13), messageLabel),
+		fyne.NewContainerWithLayout(layout.NewFixedGridLayout(fyne.NewSize(txDetailsData.MinSize().Width, txDetailsData.MinSize().Height-200)), txDetailsScrollContainer),
+		widgets.NewVSpacer(10),
+	)
+
+	txDetailsPopUp = widget.NewModalPopUp(widget.NewVBox(fyne.NewContainer(txDetailsOutput)), historyPage.Window.Canvas())
+}
